+++ bin/c.js
@@ -9,6 +9,7 @@ const { missingCommand } = require('./lib/c');
 program
     .command('d <command>', 'Shortcuts to control Docker.')
     .command('dc <command>', 'Shortcuts to control Docker Compose.')
+    .command('project <command>', 'Shortcuts to help with project management.')
     .command('hosts <command>', 'Shortcuts to help with hosts management.')
     .command('mk <command>', 'Shortcuts to control Minikube.')
     .command('npm <command>', 'Shortcuts to help with NPM.')