+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This file is a history of the changes made to @idearium/cli.
 
 - Added `c mk delete`, `c mk hosts`, `c mk ip`, `c mk start`, `c mk stop` and `c mk restart` to control Minikube.
 - Added `c hosts add`, `c hosts get` and `c hosts remove` to help with easy hosts management.
+- Added `c project env set` to help with project environment management.
 
 ## 1.0.0-alpha.12